getting rid of size_t

+++ include/fl/model/process/joint_process_model_id.hpp
@@ -127,7 +127,7 @@ public:
      *
      * Determines the joint size of the state vector of all models
      */
-    virtual constexpr size_t state_dimension() const
+    virtual constexpr int state_dimension() const
     {
         return expand_state_dimension(CreateIndexSequence<sizeof...(Models)>());
     }
@@ -137,7 +137,7 @@ public:
      *
      * Determines the joint size of the noise vector of all models
      */
-    virtual constexpr size_t noise_dimension() const
+    virtual constexpr int noise_dimension() const
     {
         return expand_noise_dimension(CreateIndexSequence<sizeof...(Models)>());
     }
@@ -147,7 +147,7 @@ public:
      *
      * Determines the joint size of the input vector of all models
      */
-    virtual constexpr size_t input_dimension() const
+    virtual constexpr int input_dimension() const
     {
         return expand_input_dimension(CreateIndexSequence<sizeof...(Models)>());
     }
@@ -163,7 +163,7 @@ protected:
 private:
     /** \cond INTERNAL */
     template <int...Indices>
-    constexpr size_t expand_state_dimension(IndexSequence<Indices...>) const
+    constexpr int expand_state_dimension(IndexSequence<Indices...>) const
     {        
         const auto& dims = { std::get<Indices>(models_).state_dimension()... };
 
@@ -174,7 +174,7 @@ private:
     }
 
     template <int...Indices>
-    constexpr size_t expand_noise_dimension(IndexSequence<Indices...>) const
+    constexpr int expand_noise_dimension(IndexSequence<Indices...>) const
     {
         const auto& dims = { std::get<Indices>(models_).noise_dimension()... };
 
@@ -185,7 +185,7 @@ private:
     }
 
     template <int...Indices>
-    constexpr size_t expand_input_dimension(IndexSequence<Indices...>) const
+    constexpr int expand_input_dimension(IndexSequence<Indices...>) const
     {
         const auto& dims = { std::get<Indices>(models_).input_dimension()... };
